Add girder_always_build_assets to Ansible role
There are situations in which assets should always be rebuilt. For
instance, local Vagrant deployments mount Girder so the first vagrant up
would result in assets not being built.

Typically for external deployments (where no such mounting exists), this
variable would be off (the default) as you would only want to rebuild
when new code has been deployed.

This also removes the 'Build Girder (web)' handler and replaces it with
a custom task. Until Ansible provides a nice way to conditionally send
notifications to handlers, this is how it has to be done.